version2/requirements.txt
absl-py==0.8.1
asn1crypto==0.24.0
astor==0.8.0
attrs==19.2.0
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0.post1
beautifulsoup4==4.7.1
bleach==3.1.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
configparser==4.0.2
contextlib2==0.6.0.post1
cryptography==2.1.4
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
enum34==1.1.6
Flask==1.0.3
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.3.0
gast==0.3.2
google-pasta==0.1.7
grpcio==1.24.1
h5py==2.10.0
idna==2.6
importlib-metadata==0.23
ipaddress==1.0.22
ipykernel==4.10.1
ipython==5.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.1.1
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==5.2.0
jupyter-core==4.6.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
Markdown==3.1.1
MarkupSafe==1.1.1
mistune==0.8.4
mock==3.0.5
more-itertools==5.0.0
nbconvert==5.6.0
nbformat==4.4.0
notebook==5.7.8
numpy==1.16.5
olefile==0.45.1
opencv-python==4.0.0.21
pandocfilters==1.4.2
pathlib2==2.3.5
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.0
prometheus-client==0.7.1
prompt-toolkit==1.0.18
protobuf==3.10.0
ptyprocess==0.6.0
pycrypto==2.6.1
Pygments==2.4.2
# pygobject==3.26.1
pyrsistent==0.15.4
pytesseract==0.2.6
python-dateutil==2.8.0
pyxdg==0.25
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.5.5
requests==2.21.0
scandir==1.10.0
scipy==1.2.2
SecretStorage==2.3.1
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.12.0
soupsieve==1.7.3
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.3
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wikipedia==1.4.0
wordcloud==1.5.0
wrapt==1.11.2
zipp==0.6.0
